harden pico cross-core control path and finish the int64 rms chain
Concurrency and control-core fixes for the pico firmware. One package, one
commit; verification per item below.

C3 — fifo pushes no longer wedge a core. All core0->core1 notifies go through
notify_core1() (push_timeout_us, drop on full); the GP9 ISR drops its refusal
on a full fifo instead of blocking. core1->core0 manual fire/peak requests are
best-effort (timeout, host can retry); arm/disarm stay blocking but only from
core1's main-loop context, which resolves within one bounded fire. Verified by
reasoning about fifo direction + drain points; bench-confirm under a stalled
host.

H3 — core 0 is the sole runtime writer of armed and arm_deadline_us. core 1
posts MAILBOX_REQ_ARM/REQ_DISARM and the disconnect handler posts REQ_DISARM
rather than clearing armed itself, closing the lost-disarm hole. The deadline is
now a single-context value so the torn 64-bit store is gone. Verified by writer
audit (grep) + reasoning; bench-confirm the disconnect-while-armed disarm.

M1 — the FIRE_IN ISR keeps the latency-critical first edge (shutter open + cam
setup busy-wait) via strobe_fire_begin(), then returns. The blocking DMA wait
and the latent sleep_ms pre-trigger delay no longer run in interrupt context;
the loop finalises the train (release XTR, stamp time, emit EVENT) once
strobe_is_idle(). Bench-measure FIRE_IN -> strobe latency to confirm it's
unchanged.

M2 — shared ADC mux serialised. FIRE_PEAK raises an in-flight flag under a
spinlock and selects its channel; core1's VSYS read backs off (returns 0) while
the peak owns the mux, and takes the same lock for its own short read so the
flag check and select can't interleave. Bench-confirm a STATUS during a sweep.

M3 — i2s_sync_head() now detects producer overrun (available > ring capacity),
skips tail to the newest full window with an even drop so the detector's L/R
phase stays aligned, and LOGs once per episode. Lives in main.c (the consumer)
rather than the unowned ring_buffer.c.

M5 — EVENT RMS streaming is gated on stdio_usb_connected() so a vanished host
can't stall core 1 on CDC backpressure. (DTR-high-but-not-draining still relies
on the connected check; flagged for bench.)

Low — STATUS line buffer sized for the worst case (32 intervals) and the
formatter reserves the newline slot, so it can't truncate or lose the trailing
newline. CMD_NONE is a silent no-op now, so a whitespace-only line stops barking
"invalid command".

H2 (from WP-A) — widened the arm-quiet RMS chain to int64 end to end: sensor
vtable current_level, sensors_max_level, mic/optical impls, fw_current_rms, the
current_rms vtable pointer, and the dispatcher's quiet-gate comparison. A loud
strike past INT32_MAX no longer wraps negative and sneaks an arm through.

Host tests: added dispatcher coverage for CMD_NONE/CMD_INVALID, the
whitespace->CMD_NONE parse, and the int64 arm-quiet gate (refuses an arm when
RMS exceeds INT32_MAX). 27 assertions pass. All four UF2s build clean.

+++ pico/include/sensor.h
@@ -12,13 +12,18 @@ typedef struct sensor {
     const char *name;
     bool    (*init)(void);
     bool    (*step)(bool armed, int32_t *level_out);
-    int32_t (*current_level)(void);
+    /* current_level is int64 because the mic's high-band RMS is a mean-square
+     * energy that overflows int32 on a loud strike (see impact_detect.h). The
+     * arm-quiet gate compares against it, so the whole getter chain stays
+     * 64-bit to avoid the negative wrap that used to let a loud room sneak an
+     * arm through. The per-step `level_out` is a separate (still int32) path. */
+    int64_t (*current_level)(void);
     void    (*selftest_append)(char *buf, int buflen, int *cursor);
 } sensor_t;
 
 bool    sensors_init_all(void);
 bool    sensors_step(bool armed, int32_t *peak_level_out);
-int32_t sensors_max_level(void);
+int64_t sensors_max_level(void);
 void    sensors_selftest_append(char *buf, int buflen, int *cursor);
 
 #ifdef __cplusplus
